Introduce JsonIndexAccessor
This is a "proof of concept" which introduces a JsonIndexAccessor as a
complement to the existing JsonPropertyAccessor.

ArrayNode indexing support has been removed from JsonPropertyAccessor
and now resides in JsonIndexAccessor.

The benefit of this set of changes is that the JsonPropertyAccessor now
focuses on property access and JSON arrays can now be indexed via integer
literals (e.g., [1]) instead of only via string literals representing
integers (e.g., ['1']).

See gh-9383